Revert to int for entity ID's

+++ VoiceCraft.Client/VoiceCraft.Client/Network/VoiceCraftClient.cs
@@ -169,7 +169,7 @@ namespace VoiceCraft.Client.Network
                 ConnectionState != ConnectionState.Connected || Muted) return;
             Array.Clear(_encodeBuffer);
             var bytesEncoded = _encoder.Encode(buffer, Constants.SamplesPerFrame, _encodeBuffer, _encodeBuffer.Length);
-            var packet = new AudioPacket((byte)_serverPeer.RemoteId, _sendTimestamp, bytesEncoded, _encodeBuffer);
+            var packet = new AudioPacket(_serverPeer.RemoteId, _sendTimestamp, bytesEncoded, _encodeBuffer);
             SendPacket(packet);
         }
